+++ src/markdown_parser.py
@@ -130,7 +130,7 @@ def _add_block_format(
             if block_type == "paragraph":
                 block = blocks.paragraph(
                             content     = string,
-                            annotations = blocks.create_annotations(
+                            annotations = blocks.add_annotations(
                                 bold = bold,
                                 italic = italic,
                                 strikethrough = strikethrough,
@@ -141,7 +141,7 @@ def _add_block_format(
                 block = blocks.to_do(
                             checked     = special_field,
                             content     = string,
-                            annotations = blocks.create_annotations(
+                            annotations = blocks.add_annotations(
                                 bold = bold,
                                 italic = italic,
                                 strikethrough = strikethrough,
@@ -152,7 +152,7 @@ def _add_block_format(
                 block = blocks.heading(
                             heading_num = special_field,
                             content     = string,
-                            annotations = blocks.create_annotations(
+                            annotations = blocks.add_annotations(
                                 bold = bold,
                                 italic = italic,
                                 strikethrough = strikethrough,
@@ -162,7 +162,7 @@ def _add_block_format(
             elif block_type == "toggle":
                 block = blocks.toggle(
                             content     = string,
-                            annotations = blocks.create_annotations(
+                            annotations = blocks.add_annotations(
                                 bold = bold,
                                 italic = italic,
                                 strikethrough = strikethrough,
@@ -172,7 +172,7 @@ def _add_block_format(
             elif block_type == "bulleted_list_item":
                 block = blocks.bulleted_list_item(
                             content     = string,
-                            annotations = blocks.create_annotations(
+                            annotations = blocks.add_annotations(
                                 bold = bold,
                                 italic = italic,
                                 strikethrough = strikethrough,
@@ -182,7 +182,7 @@ def _add_block_format(
             elif block_type == "quote":
                 block = blocks.quote(
                             content     = string,
-                            annotations = blocks.create_annotations(
+                            annotations = blocks.add_annotations(
                                 bold = bold,
                                 italic = italic,
                                 strikethrough = strikethrough,
@@ -192,10 +192,10 @@ def _add_block_format(
 
         # Append to paragraph block text with specific annotations if any
         else:
-            blocks.append_text_to_block(
+            blocks.append_rich_text(
                 notion_block    = block,
                 content         = string,
-                annotations     = blocks.create_annotations(
+                annotations     = blocks.add_annotations(
                                     bold = bold,
                                     italic = italic,
                                     strikethrough = strikethrough,